proactive-python-client-examples-ai/machine-learning/logistic-regression/requirements.txt
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
# optional (to be used with proactive)
py4j>=0.10.9.7
python-dotenv==1.0.1